+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Notes
   - Implemented pagination, limiting the initial display to 100 items and loading additional history items as you scroll.
 - **Improve icon display on macOS Golden Gate**
 - **Improve the add apps to exclude screen**
+- **Improve the preset picker display on macOS Tahoe or later**
 - **Changes to the preset icon picker package**
   - Icon names are now displayed, and you can filter by specific categories using the category picker.